Start splitting up the webui
