+++ src/atomic.py
@@ -147,18 +147,3 @@ def ros_is_running():
 	if (rospy.is_shutdown()):
 		return TruthValue(0, 1)
 	return TruthValue(1, 1)
-
-# Update dynamic paramater cache
-def update_opencog_control_parameter(name_node, value_node):
-	try:
-		name = name_node.name
-		value = float(value_node.name)
-		evl.update_opencog_control_parameter(name, value)
-		return TruthValue(1, 1)
-	except:
-		return TruthValue(0, 1)
-
-# Update dynamic parameters
-def push_parameter_update():
-	evl.push_parameter_update()
-	return TruthValue(1, 1)
