+++ webui/README.md
@@ -1,0 +1,5 @@
+
+WebUI Control panel
+===================
+
+Stuff for controlling the OpenPsi paramaters via a web user interface.
